pyproject.toml
[tool.poetry]
name = "fitai"
version = "1.0"
description = "Fit AI is a bot for creating courses with custom settings"
authors = ["Undo <github.com/UndoGG>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
tortoise-orm = "^0.20.0"
aiogram = "^3.4.1"
pydantic = "2.4.1"
rich = "^13.7.1"
click = "^8.1.7"
aiohttp = "^3.9.3"
pyyaml = "^6.0.1"
python-dotenv = "^1.0.1"
fastapi = "^0.110.0"
uvicorn = "^0.28.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
